Add request_params and handle redirects if allow_redirects disabled
With request_params, we can include any parameters that are available
for aiohttp.ClientSession.request. In this patch, we handle redirects
manually if allow_redirects is set to False.

We also added a force_http attribute to the AsyncSession class. This
attribute is used to change a redirect URL's protocol from https to
http, if necessary.

Bug: T363725